Adding a temporary hack for zipping up all plugins (agent and server)
This is temporary work around for hudson builds. Zipping up all of the
plugins is normally done as a post-build action using the M2 Extra Steps
hudson plugin; however, that plugin is not yet available on the new
hudson server. Once the plugin is available, the changes in this commit
will be undone.