[BZ 699542 - Intermittent Serialization issues with Hibernate types]
This is a fix that I think solves the issue. It introduces changes
and logging that indicate that a potential problem in HibernateDetcahUtility
has been identified and avoided. In short we treated System.identityHashCode()
as if it guaranteed unique values, which it may not.  We now handle the
situation whete the identity hash may in fact not be unique among the set
of objects being scrubbed.

The logging will be tuned down in a subsequent commit but for now dumps out
the objects that have the same hash but are not the same.