First cut and copying and compressing the snapshot data
This is a first cut at copying the files to be included in the snapshot.
The files are zipped up and stored in the Snapshot as a byte array.
SnapshotGeneratorTest has been updated to verify this works as expected
(so far).

There are a couple enhancements that will need to be considered at some
point. 1) What is the best compression algorithm to use? I am using ZIP
for now primarily because it is already used in the code base and so was
easy to get working. Other compression strategies though may be faster
and/or produce smaller files. 2) Should the data be stored directly in a
byte array or accessed lazily through a stream to avoid the memory
overhead?