Refactor SnapshotGenerator to write meta data to disk
The snapshot meta data is now written to disk and not stored in memory.
Meta data could become rather large so we need to be careful about
when/where we fully realize it in memory. The generateSnapshot method
now returns a SnapshotHandle object which is basically a pointer to the
snapshot meta data and data files on disk. Unit tests have been
refactored as well for these changes.