Initial commit for several classes needed for agent/server drift flows
The classes at this point are more or less stubs. DriftManager is the
agent service that will handle drift operations. DriftServerService is
the server-side API that agents will invoke. It currently defines a
single method, uploadSnapshotReport(SnapshotReport report).
SnapshotReport contains input streams for both the meta data and data
files that are suitable for use in remote method calls.

This is the initial commit for the DriftManagerBean SLSB. All it does
for now with uploadSnapshotReport method is write the bits to a local
directory.

A getDriftManager method has been added to LookupUtil.