Get DriftServerService working so agent can upload snapshots
This commit includes a number of changes. First, the
rhq.communications.remote-pojos property in server-comm-configuration.xml had to
be updated to include DriftServerService so that it can receive remote requests.

DriftServerService.uploadSnapshot no longer takes a SnapshotReport argument. The
input streams for the meta data and data files are properties of SnapshotReport.
The server was not able to consume these streams. The comm layer can do
remote streaming, but it needs to decorate each stream. The comm layer only
looks at each top-level parameter and if it is a stream it will decorate it so
that it can be consumed remotely. My streams though included as properties of
SnapshotReport and thus were not getting decorated. I have changed the signature
of uploadSnapshot() so that the input streams are now top level
parameters.