[BZ 703252 - IE (IE7, IE8) unable to load CoreGUI]
Revert the use of DOCTYPE as "standards" mode causes various issues in
several browsers.  This is basically due to the fact that SmartGWT codes
for browser's native "quirks" behavior.  As such SmartGWT is more
predictable when avoiding "standards" modes.  But we still had a need to
direct IE in some way to give us the behavior we want. Todo this we add the

  <meta http-equiv="X-UA-Compatible" content="IE=EmulateIE7" >

directive to our initializing html (CoreGUI.html).  This instructs IE to
emulate IE7 quirks mode.  In effect, this means that if our code looks good
with IE7 it should continue to work well even if users upgrade to IE8, IE9,
etc...  It also seems to provide enough guidance to IE to get perform the
proper loading, assuming we build the "ie6" gwt.userAgent.  And it gets
around a SmartGWT tree rendering issue with IE8 modes.

Since this is ignored by non-IE browsers they will continue to run in their
own quirks mode, thus avoiding issues like BZ 703515, which is a regression
brought about by the original DOCTYPE approach.


Alternatively,

Conflicts:

	modules/enterprise/gui/coregui/src/main/webapp/CoreGUI.html