[BZ 536032 - (RHQ-426) session oddities when the same user is logged in from multiple machines]
Historically a username shared an RHQ session (not the HTTP session, the
underlying RHQ session, the sessionId on the Subject).  This had some mild
upside when it was convenient to share a session but has serious downside
as the logout of one gui session or cli job would invalidate the session
for any other gui or cli session using the same username.

So, this changes the behavior to make things behave more in line with what
folks hopefully expect. Each login gets its own RHQ session. Logout or
session timeout will not affect other logins.

One exception may be multiple sessions in the same browser, as there could
be conflict in cookie handling. This is not new and not a use case we are
planning to address.

In addition, I replaced some SubjectManagerLocal methods and supporting
queries with equivalent Criteria calls.