The server jar (EJB3) tests did not run because the abstract EJB3 test container could not bootstrap:
Caused by: java.lang.NoSuchMethodError: org.jboss.ejb3.KernelAbstractionFactory.setKernel(Lorg/jboss/kernel/Kernel;)V

This was due to the fact that the embeddable EJB3 jar was declared below
the standard EJB3 jars in the pom's depedencies. And the class was getting loaded
from the wrong jar.  The maven classpaths are ordered in the order of
deps defined in the pom.  I've moved the test scoped deps above the
non-test 3rd party deps.  This should not affect non-test builds as the test deps
should be ignored in those builds.
Also:
- removed a duplicate dep for commons-logging in the server jar pom
- changed some deprecated ${version} and ${groupId} declarations to
  the supported ${project.version} and ${project.groupId} decls, respectively.



deps