Uploading snapshots is now done asynchronous with an MDB
Initial commit for UploadSnapshotBean and UploadRequest.
UploadSnapshotBean is an MDB that now handles the uploading of snapshot
files. DriftManagerBean has been refactored so that it no longer handles
the upload and instead just sends a message to the jms queue for
UploadRequestBean to consume. When an agent calls the server to upload a
snapshot, the thread that handles the request will now block for a
minimal amount of time, freeing up the thread to service other agent
requests or UI requests.