[Bug 700121] LDAP Group Member search doesn't escape special characters
Added method encodeForFilter to LdapGroupManagerBean. This allows filter criteria to be properly encoded when the search filter is constructed.
encodeForFilter method is used when adding the userDN to the search filter.
Cleaned up some log messages to be easier to identify what was actually occurring at the time the message was logged.
Added additional debug log messages to help identify what criteria is being used to perform LDAP searches.