Increasing the number and size of snapshots to verify lazy loading
I increased the number of snapshots to 25 with data files of 100 MB
which comes in at about 2.4 GB of data. When the test fetches the
Snapshots, it stores them in a list so they are held onto for the
duration of the test. If those blob fields were not getting lazy loaded,
at 2.4 GB of data, I have to think that an OutOfMemeoryError would be
getting thrown.