Updating tests to avoid transaction timeouts
loadMultipleSnapshotsWithoutLoadingData() was loading all of the
snapshots in a single transaction which was timing out. This isn't
surprising giving the size and number of snapshots. I have refactored
the test to persist and load each snapshot in a separate transaction.

Adding the test updateSnapshotData(). This method first persists a new
snapshot, then loads it from the database, then updates the blob, and
then finally writes it back to the database.