[BZ 707759 - Deployment of new WAR to EAP server fails due to thread timeout - Need configurable thread timeout] Provide overrides for the default 60 second timeout of create and delete child actions.
In general 60s may be enough but for applications that need time to deploy,
or perhaps to shut down gracefully, this can be too short.  The overrides
are blunt in that they affect all types. And it should be noted that while
creating(deploying) or deleting the child resource, the parent resource will
be write-locked, meaning deferred metric reporting etc. So, it it recommended
that when possible fast create and delete is performed for these facet
implementations.