The original check-in done via the Eclipse resource bundle tool caused issues in the resource bundles using native unicode multibyte characters Specifically, the JA and DE locales.  The native unicode was converted to escaped unicode (\uZZZZ format) in a byte-wise fashion. There were two problems, first, the plugin had an option set to perform the conversion (this should be unchecked) and the file encoding was "ISO-8859-1". According to the GWT I18N docs, it should be "UTF-8".
This check-in returns JA and DE to using native unicode. ZH and PT remain
encoded as that is how the translators submitted the files (it could go
to native if they desire to do that conversion).  Secondly the bundles and
constants files are now declared as "UTF-8" encoded.

This is an Eclipse preference setting, it may be that IJ will need to make\
a similar declaration.

Hopefully this settings going forward. The wiki will be updated.