Initial commit for DriftMetadataParser, refactoring plugin parser tests
DriftMetadataParser handles creating the underlying configuration
definition as well as the default template. Since the drift schema
requires some values to be set, the plugin parser has to generate a
default template in addition to the configuration definition it
generates.

I have refactored tests in PluginMetadataParser to verify that the
correct property definitions are generated as well as the correct
properties in the default template configuration.