Adding dbsetup/dbupgrade to generate the global drift config def
I was previously generating a config def in the plugin parser code for
each resource type that declares one or more <drift> elements. There is
no reason though to generate and persist those config defs because the
drift configuration definition is defined in rhq-drift.xsd.

This all became clearer to me after collaboration with jshaugn. The
plugin parser needs to be generating config templates which will be
handled in a subsequent commit.