Adding support for drift templates in the data model
A drift template (or drift configuration template) corresponds to a
<drift> element declared in the plugin descriptor. A <drift> element is
parsed into and represented as a config template. A resource type can
have multiple drift templates.

This commit includes the dbsetup/dbupgrade work for this change. It also
includes the JPA mappings and unit test to verify the correctness of the
mapping. Note that ResourceTypeTest is disabled in domain/pom.xml. I
think it is disabled because I implemented that test with unitils and if
memory serves, there were problems running unitils based tests along
side our embedded container tests. You should however be able to run
ResourceTypeTest from your IDE or from maven quickly and easily by
commenting out the dependsOnGroup attributes.