Removing ResourceType association from ConfigurationTemplate
I initially defined the mapping in ConfigurationTemplate because that
was the only way I knew how to configure the one-to-many with a mapping
table. In addition to not really needing ResourceType in
ConfigurationTemplate, it also caused a lot of regressions in tests. I
figured out how to completely drive the assocation from ResourceType.