Initial commit for a few classes/interfaces involved with drift detection
DriftDetecor will run in its own thread at scheduled intervals. It gets
the next drift detection schedule from the schedule queue and then will
start with the file system scan. The schedule queue is a priority queue,
and the schedules are ordered based on timestamp such that the oldest
schedule will move to the top of the queue.

This is the initial commit for the ChangeSetWriter classes. The
changeset writer will be responsible for generating the meta data file
that stores SHAs and paths as well as for copying the content that is
included in the changeset.