Changing DriftAgentService.requestDriftFile to take a Headers arg
The headers from the change set meta data file is what is really needed
to locate the drift files. Like the drift config, it contains the drift
config name and basedir, but it will also contain a version number that
may be needed. That version number is not stored in the drift config.

Refactoring ChangeSetManager APIs to only require what is needed for
arguments.