Adding support for sending change set content to server
Initial commit for DriftFilesSenderTest and DriftClient. DriftClient is
an interface that is passed into DriftDetector and DriftFilesSender to
better facilitate testing.

This is a rough, first cut at this functionality that has not yet gone
through end-to-end testing. DriftFilesSender iterates through the change
set meta data file for each DriftFile until it finds a matching hash.
When a match is found, it stages the file so that it can be sent to the
server.