First, rough cut at generating drift change sets
Some of the code is particularly rough because it is going to be
refactored soon as a new change set file format is introduced. The file
format will be more similar to that used by bundles where files are not
grouped by directories. I am trying not to do any more than I have to in
order to get things working so that I can start the refactoring.

This commit also adds logic in DriftDetector to check whether the
schedule pulled off the queue is ready to be processed. If not, it is
put back on the queue, and DriftDetector simply returns.

This commit does not include logic for updating the coverage change set
when drift is detected.