Fixing IO error when streaming drift files to the database
No longer using ZipUtil.walkZipFile because an IOException was getting thrown
after reading the first entry, resulting in subsequent entries being skipped.
DriftFileVisitor passed the ZipInputStream to Hibernate.createBlob, and either
Hibernate, the JDBC driver, or something else is closing the stream which in
turn causes the exception.

Now we unzip the content file into a temp directory and pass a
FileInputStream to hibernate for each file in the directory. Once each
file has been successfully stored, we delete the directory.