[BZ 713744] Fixing another deadlock issue that can occur during plugin installation
This deadlock was discovered and reproduced by running the test
UpdateConfigurationSubsystemTest.testAddDeleteTemplate. The deadlock
occurs in ResourceMetadataManagerBean.mergeExistingType on line 397
which is,

  resourceConfigMetadataMgr.updateResourceConfigurationDefinition(existingType,
      resourceType);

The problem as with all of these deadlock issues involves the table
rhq_config. In the test method, a plugin that declares a config
template, is installed and then updated. In an initial transaction that
is started in PluginManagerBean, a write lock is allocated to update
rhq_plugin. updateResourceConfigurationDefinition executes in its own
transaction, and because the test involves some config templates, a
write lock is needed for rhq_config since there is a one-to-one
assocation between rhq_config_template and rhq_config. That lock cannot
be obtained because rhq_plugin has a FK with rhq_config for the
ServerPlugin class, hence the deadlock.

With this commit, PluginManagerBean.registerPlugin has been refactored
to avoid the deadlock. It first calls a new method, installPluginJar,
which executes in its own transaction and handles updating the Plugin
object. After that method is called, registerPluginTypes is called. That
call starts a new transaction, and registerPluginTypes now only handles
the meta data updates.