Fixing a bug for deleting config templates during plugin installation
This issue came to light as a result of
ConfigurationUpdateSubsystemTest.testAddDeleteTemplate failing with the
following exception,

javax.persistence.EntityNotFoundException: deleted entity passed to persist:
[org.rhq.core.domain.configuration.definition.ConfigurationTemplate#<null>]
...
at
org.rhq.enterprise.server.configuration.metadata.ConfigurationMetadataManagerBean.updateTemplate(ConfigurationMetadataManagerBean.java:270)
at
org.rhq.enterprise.server.configuration.metadata.ConfigurationMetadataManagerBean.updateConfigurationDefinition(ConfigurationMetadataManagerBean.java:181)

In testAddDeleteTemplat a plugin was instaled that defined a couple
templates in the resource configuration. The plugin is upgraded, and
the newer version of the plugin removes one of the templates. In
ConfigurationMetadataManagerBean, we call EntityManager.remove on the
template that is to be removed, and then we proceed to update the other
template. Then we call EntityManager.flush while the deleted template is
still attached to the parent config def. This causes the exception.

I have no idea why we were not seeing this before now, but this commit
resolves the issue. We now first identify which templates are to be
removed and which are to be updated. Then we call EntityManager.remove
on those templates to be removed, and we also remove them from the
owning config def. Finally, we proceed to update the other templates.