Store @Id field as a String while storing it as an int in the database
Persistence management of drift-related entities will be managed by
server plugins. The intent is to at some point implement one or more
plugins that scale better (horizontally) than the database. Relying on
the database to generate ids as we do with the database can be
problematic in a clustered environment. It is easier to use something
like UUIDs which can generated by client code. UUIDs can easily be
represented as strings so we want to make the id fields Strings. At the
same time, we do not want to comprimise the convention and advantages of
using auto-incrementing ids as is done with the single database.

To get the best of both worlds, I am introducing a hibernate custom
type, IntegerString, which allows us to store the id as a string and
convert to an integer in the database.