Make sure files exist when the server requests agent to send them
DriftFilesSender was not checking that files existed before attempting
to copying them to the staging area to be zipped up and sent to the
server. This resulted in an exception getting thrown, preventing other
files from getting sent to the server in that run of DriftFilesSender.
A check has been added to avoid this.