Adding config_id to rhq_drift_change_set table and to DriftChangeSet interface
We need to store the drift config id in the change set table so that we
can identify change sets. A change set belongs to a drift configuration
and a resource can have multiple drift configurations; so we need both
the resource id and the config id to identify change sets that map to a
particular drift config.