Initial commit for DriftComposite and removing Resource reference from DriftChangeSet
The Resource reference has been removed from DriftChangeSet because
drift server plugins are responsible for managing the persistence of
drift-related entities. The get/setResource methods in DriftChangeSet
have been replaced with get/setResourceId.

DriftDataSource in the gwt code was calling the getResource() method
that has been removed. DriftComposite has been introduced to deal with
the change. DriftDataSource now fetchs a list of DriftComposites which
include both the Drift and owning Resource objects.