Initial commit for JPAUtils
Refactoring some methods in AbstractEJB3Test to delegate to JPAUtils
with the intent to eliminate duplicate code in core/domain and
server/jar tests. Most importantly though, the method clearDB is now
part of JPAUtils making it accessible to any test. This provides a much
more robust way of resetting the entire database to a known, consistent
state.