Adding support for updating drift configs
When the server sends a request to the agent to update a drift
configuration, that config is attached to a schedule which is in one of
two places. It is either on the schedule queue or it is "active"
schedule currently being processed by the drift detector. The config
update will be applied such that the next time the drift detector
processes the schedule the update will take effect.