Adding support in coregui for string ids
TableSection has a lot of code driven off of integer ids, and there are
a lots of subclasses. I started out trying to refactor it to support
string ids, but due to concerns over introducing regressions in existing
(and working) subclasses, I decided to introduce TableSection2 which
deals with string ids. After review, this code may get refactored or
altogether go away with the logic being merged into TableSection.