fixing failing test
The test failure was caused as a result of a code change in
DriftManagerBean, and the corresponding test code was not updated. The
logic for setting change set version numbers was updated to assign
unique version numbers across drift configs (and their owning resource)
whereas previously version numbers were not unique for change sets
belonging to a particular drift config.