starting to add support for filtering during drift detection
This commit adds the filtering login in FileUtil. I put it there
because it is general purpose code that could be used in placed other
than drift code. As part of that I pulled the Filter class out of
DriftConfiguration and moved it into core/util for use with the
filtering.