Fast Click Problem Work Fast clicking in the UI has caused various exceptions. We think it may be due, in part, to a bad assumption made in various parts of the code. And that is that the current history token (History.getToken()) has not changed since the widget has been instantiated.  If the widget is still initializing while the user navigates away the assumption may be invalid.
This is the first of at least two checkins removing the bad approach. This one
works solely on changing portlet creation, which relied heavily on this
technique, specifically in the resource and group portlets in the Activity
subtabs.

The general approach is to push more information into the constructors/factories
in the shape of EntityContext objects.  The EntityContext itself has been
beefed up slightly to be able to represent better AutoGroup and AutoCluster
groups, as well as had a few convenience methods added.

Also:
- Fixed a problem in Drift Facet/Subtab handling
- Removed the assumption from portlets that bundles can only be deployed to
platforms.
  -- Added new Bundle Facet