Fast click resource tree nav work It remains to be seen if this gives us coverage from the problem but the main issue I found, and have tried to resolve is the following: After clicking a node in the tree, and prior to that detail view being able to establish its subtab views, the user navigates to another node in the tree.  Navigating between resource and autogroup nodes seems to exacerbate the problem.
I've identified the point in the rendering of the detail view where, once
completed, it should be safe to click in the tree. There were two approaches
I tried locally to solve the issue.  The first was to catch and swallow
exceptions thrown due to quick nav.  This worked well with one exception, it
cause some widgets to fail being destroyed, get detached and then render in
the top left corner of the GUI. I could not find a good way to lazily clean
that up. The second, implemented here, is preventative.  Disable the tree
until we get past the safety point in rendering the detail view.

Whether this is irritatig or not remains to be seen, but at the moment
I don't see an alternative nor does the disable period seem obtrusive.
This is accomplished by using a listener approach, where the Top view
registers a listener with the detail view, and re-enables the tree when
the detail view says it has rendered.