Adding initial support for generating a unified diff
This commit adds support in the SLSB layer for generating a unified diff
for a Drift object. It compare the two versions referenced by the drift
object.

This commit also adds the ability to view file contents on
DriftDetailsView. This may very well be throw away work, but it giving
me more practice in working with SmartGWT. And the next step is to add
the ability to display a unified diff in the UI. Then I will have the
ability to view different versions of a file as well as generate a diff
of them.