Do not have the change set reader automatically close stream when reading  change set file
The reader was automatically closing the stream which caused an
exception in ZipUtil.walkZipFile. This didn't cause a loss in
functionality since the zip file only has one file; however this fix is
needed to avoid noise in the logs and also to handle the future
possibility of the change set zip containing multiple files.