Check to see if a schedule has been removed before sending change set to server
In my previous commit I added support for deleting a drift detection
schedule and purging the change set directory. If the deletion occurs
while the schedule is still active, DriftDetector finishes its run and
the agent could still wind up sending a change set report to the server
for a configuration that has been deleted.

This commit adds a check in DriftManager.sendChangeSetToServer to make
sure that the schedule has not been deleted before sending the change
set to the server.