[BZ 732078] Adding support for detecting "viewable" files
This commit introduces a simple approach for determing whether or not a
file is viewable. The approach taken is try and determine whether or not
the file is a binary file. DriftManagerBean uses a hard coded list of
known file extensions for binary files. This is in no way as robust as
it could be but it may be sufficient. Since the solution is not robust,
I think it is important to try and detect if a file is a binary type as
opposed to detecting whether or not it is a text file suitable for
viewing. The rationale behind that is we do not want to prevent a user
from viewing a file or viewing a diff of a file when they should be able
to.