[BZ 733144 - links to non-default subtabs don't work] The problem occurred, I believe, when you navigated from resource A, default tab/subtab to resource A, non-default tab/subtab.
An easy example of this would be 1) Favorites Button->resources->A,
bringing you to A->Summary-Activity and then 2) Click on the alert def
name for a recent alert in the alerts portlet.

This scenario manged to generate a smartgwt tabbing event that did not play
well with others.  As a response, I've expanded the net ignoring certain
smatgwt tabbing events to catch this as well.  Tree/Tab/Subtab
nav should be well exercised (tree/context menu/tabs/subtabs) to ensure
thia change solves the problem without causing others.