[BZ 734194] Check that content is loaded in db on drift details view
The drift details view page displays links for viewing the current and
previous version of a file as well as a diff between those versions.
The initial implementation failed to take into account that the file
content referenced by the drift may not be in the database. The page
previously attempted to generate a diff report on load instead of on
demand. This of course resulted in an exception getting thrown.

With this commit, links for viewing file content and for viewing a diff
are shown when the file content is available. If the file content is not
available, a message is instead displayed that says the content is not
available.