[BZ 734881] Need to check for empty content because oracle blobs don't handle it
When a file is empty we still persist a row for it in rhq_drift_file.
The blob is set to null. This works fine on postgresql, but results in
the NPE reported in BZ 734881. To work around this, we now check the
size which is also stored in the rhq_drift_file table, and if the size
is zero, DriftManagerBean.getDriftFileContent just returns an empty
stream instead of opening an input stream on the blob.

This required one additional change. We previously were not storing the
size in rhq_drift_file. I have updated
JPADriftServerBean.persistDriftFileData to save the size now as well.