Override default tmp directory used during gwt compilation
Currently the GWT comopiler generates 25 GB of temporary files which by
default are written to the system temp directory. 25 GB exceeds the
default size limits for the volume on wnich the system temp directory
resides. This change causes the gwt compiler to write temp data to the
same volume in which the build is running.