Update the resource container when a drift config is deleted
Prior to this commit if you deleted a drift configuration and then
subsequently restarted your agent, the drift configuration comes back to
life on the agent. That is because it is persisted to disk along with
other resource data when the agent is shut down. With this commit, when
a drift configuration is deleted, it is now also removed from the
resource container.

ResourceContainer now also stores DriftConfiguration
objects instead of DriftDetectionSchedules. The reason for this change
has to due with inventory sync. During a sync the server will be sending
the drift configs and not schedules. The drift detection scheudles exist
only on the agent.