Detect deleted drift configs during inventory sync
This commit adds support for detecting deleted drift configs during
inventory sync. when a deleted config is detected, the change set
directory is purged.

This commit also adds supports for handling renamed configs which may be
a common scenario. Suppose for example I create a config with a set of
filters. After some drift detection, I realize the filters are not what
I want. The agent goes down, I delete the config, create a new one with
the same name but with different filters. The agent properly detects
that the original config was deleted, purges the change set directory,
and treats the new config as such, i.e., new.

The code in this commit and in the previous one are bit rough and need
some refactoring and unit tests. That is coming next...