Do not purge change set directories for configs that have not been deleted
Prior to this commit, DriftManager was incorrectly purging change set
directories at start up. This resulted in a new coverage change set
being generated. Now at start up and during inventory sync, when we
detect that a config has been deleted and only when it has been deleted,
we purge its change set directory. There are a couple reasons we purge
those directories. First, we do not want to clutter up the managed
system with old, unused change set data. Secondly, and more importantly,
the deletion has to be done to support renamed configs. A user can
delete a config, and the create a new one with the same name. This might
happen when changing filters for instance. Even though the configs have
the same names, they are different, and the new config cannot use the
old change set data.