Filter queue by resource id when checking for deleted configs
When resources are discovered after the agent has already been running,
an inventory sync is triggered, and for syncing drift configs this has a
different execution path than an inventory sync at start up because
DriftManager is already initialized. In this execution path I was
detecting deleted configs incorrectly. When comparings the configs from
the server against the configs in the schedule queue, we need to filter
the queue by resource id.