Initial commit for FileDAO and FileDAO test
The work in this commit mainly involves getting more familiar with the
GridFS which to my pleasant suprise is very easy to use. So far FileDAO
has a finder method for fething a file by id. That may be the only
finder method needed. And FileDAO test not only tests the finder but
also demonstrates writing a file to MongoDB using GridFS.