Have the agent track and store the change set version number
DriftDetector now updates the version header in the snapshot file and in
the drift change set file. The server now gets the change set version
that it stores in the db from the headers.

This change will allow the server to determine whether or not the agent
and server are in sync.