Updating MongoDB drift plugin to store file contents.
Now when a change set is reported by the agent, the mongodb plugin
checks the database to see if the content is already there. If it isn't,
it sends a request to the agent for the content. When the agent sends
the content, the plugin saves the bits to the database using GridFS.