Added support for using the import configuration specified in the export file during import.
To achieve that the export file schema now includes a new
"configuration-instance" schema, which is a combination of the existing
configuration definition schema and an ability to specify values
conforming to that definition.

Added unit tests and refactored the constants from ExportingInputStream
into a separate file so that they are in a more logical place.