Adding error handling logic on the agent during drift detection
This commit introduces several changes and a changed work flow to
address some boundary conditions that can arise when the server fails
to receive a change set report. The issue stems from the way we stream
the change set report to the server. Because the request is processed
aysnchronously, we cannot know for certainty if/when errors arise in the
comm layer.

When DriftDetector runs, a new snapshot file is generated, and now a
copy of the previous version snapshot is maintained as well. After the
server processes the change set, it now sends an ack to the agent. This
lets the agent know that the change set was successfully persisted on
the server. The agent then cleans up, deleting the previous version
snapshot, and the change set zip file.

If drift detection runs again before the agent receives the
ackowlegement, drift detection is skipped. The most likely scenario for
not receiving an acknowledgement would be a network error or a down
server.

If any errors occur during drift detection, which includes sending the
change set to the server, the agent will attempt to revert back to the
previous version snapshot. This is to ensure we have a consistent
snapshot on disk with which to work.

This commit also fixes a bug in the drift inventory sync code. In
situations where there are existing change sets on the server, and the
agent has to fetch a snapshot from the server, the snapshot version was
getting set incorrectly. This is because the snapshot was not being
built correctly. Change sets were being applied out of order. This is
fixed now.