[BZ 727959] Check to make sure files are readable during drift detection
When generating the initial change set, if a file is not readable, it is
skipped and excluded from the change set since we cannot compute the SHA
for the file.

When generating a drift change set, if we detect a new file that is not
readable, we skip it as well since we cannot compute the SHA. Now if an
existing file that is already being monitored for drift has its
permissions changed such that it is no longer readable, then it is
treated and recorded as a deleted file entry.