First, (very) rough cut of support for pinned snapshots
Some of this code, particylarly on the server may get changed
substantially or altogether go away. The key changes are primarily on
the agent side for supporting the notion of a pinned snapshot. The agent
checks to see if it is using a pinned snapshot. If so, it does not
update that snapshot file, and it will continue using the pinned
snapshot for comparison against the file system during drift detection.

There is also a new DriftAgentService method, pinSnapshot. The server
calls this to pin a snapshot for a particular drift definition. A
DriftSnapshot object is passed to the agent, and that object is written
to disk in the standard snapshot/changeset format.